Backend-Services/Consumer/requirements.txt
python-dotenv
paho-mqtt
supabase==2.7.4